+++ src/display.rs
@@ -121,8 +121,8 @@ impl<'a> Display<'a> {
         run_instr: &RunInstr,
         debug: &DebugManager,
         opcode: u8,
-        t_cycles: u32,
-        m_cycles: u32,
+        t_cycles: u64,
+        m_cycles: u64,
         halted: bool,
     ) {
         self.debug_ui.machine_state_update(
